+++ block-storage/src/test/scala/coop/rchain/blockstorage/dag/IndexedBlockDagStorage.scala
@@ -1,10 +1,11 @@
-package coop.rchain.blockstorage
+package coop.rchain.blockstorage.dag
 
 import cats.Monad
 import cats.effect.Concurrent
 import cats.effect.concurrent.{Ref, Semaphore}
 import cats.implicits._
 import com.google.protobuf.ByteString
+import coop.rchain.blockstorage.BlockStorageMetricsSource
 import coop.rchain.casper.protocol.BlockMessage
 import coop.rchain.crypto.hash.Blake2b256
 import coop.rchain.metrics.Metrics.Source
