+++ casper/src/main/scala/coop/rchain/casper/util/DagOperations.scala
@@ -3,7 +3,7 @@ package coop.rchain.casper.util
 import cats.{Eval, Monad}
 import cats.implicits._
 import com.google.protobuf.ByteString
-import coop.rchain.blockstorage.BlockDagRepresentation
+import coop.rchain.blockstorage.dag.BlockDagRepresentation
 import coop.rchain.models.BlockMetadata
 import coop.rchain.shared.StreamT
