+++ casper/src/test/scala/coop/rchain/casper/CliqueOracleTest.scala
@@ -2,8 +2,7 @@ package coop.rchain.casper
 
 import scala.collection.immutable.HashMap
 import scala.collection.mutable
-
-import coop.rchain.blockstorage.{BlockStore, IndexedBlockDagStorage}
+import coop.rchain.blockstorage.BlockStore
 import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.casper.helper.BlockGenerator._
 import coop.rchain.casper.helper.BlockUtil.generateValidator
@@ -12,6 +11,7 @@ import coop.rchain.casper.protocol.{BlockMessage, Bond}
 import coop.rchain.models.Validator.Validator
 import coop.rchain.p2p.EffectsTestInstances.LogStub
 import com.google.protobuf.ByteString
+import coop.rchain.blockstorage.dag.IndexedBlockDagStorage
 import monix.eval.Task
 import org.scalatest.{FlatSpec, Matchers}
