+++ casper/src/test/scala/coop/rchain/casper/ValidateTest.scala
@@ -5,7 +5,8 @@ import java.nio.file.Files
 import cats.Monad
 import cats.implicits._
 import com.google.protobuf.ByteString
-import coop.rchain.blockstorage.{BlockStore, IndexedBlockDagStorage}
+import coop.rchain.blockstorage.BlockStore
+import coop.rchain.blockstorage.dag.IndexedBlockDagStorage
 import coop.rchain.casper.helper.BlockGenerator._
 import coop.rchain.casper.helper.BlockUtil.generateValidator
 import coop.rchain.casper.helper.{BlockDagStorageFixture, BlockGenerator}
