+++ casper/src/test/scala/coop/rchain/casper/helper/NoOpsCasperEffect.scala
@@ -4,8 +4,9 @@ import cats.effect.{Resource, Sync}
 import cats.implicits._
 import cats.{Applicative, Monad}
 import com.google.protobuf.ByteString
-import coop.rchain.blockstorage.BlockDagStorage.DeployId
-import coop.rchain.blockstorage.{BlockDagRepresentation, BlockDagStorage, BlockStore}
+import coop.rchain.blockstorage.dag.{BlockDagRepresentation, BlockDagStorage}
+import coop.rchain.blockstorage.dag.BlockDagStorage.DeployId
+import coop.rchain.blockstorage.BlockStore
 import coop.rchain.casper._
 import coop.rchain.casper.DeployError
 import coop.rchain.casper.protocol.{BlockMessage, DeployData}
